[cocas] Revert target-specific write_image, generate better image files
