--- cocas/write_image.py
@@ -1,10 +1,0 @@
-class WriteImageInterface:
-    """
-    Write the contents or array into file in logisim-compatible format
-
-    :param filename: Path to output file
-    :param arr: Array to be written
-    """
-    @staticmethod
-    def write_image(filename: str, arr: bytearray):
-        pass
